pyproject.toml
[project]
name = "bat-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "GladeOwl"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "hikari (==2.4.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "hikari-lightbulb (>=3.2.1,<4.0.0)",
    "hikari-miru (>=4.2.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
